pyproject.toml
[tool.poetry]
name = "chloe"
version = "0.1.0"
description = ""
authors = ["anonymous"]

[tool.poetry.dependencies]
python = "^3.7"
jupyterlab = "^1.1"
pandas = "^0.25.2"
boto3 = "^1.10"
orion = "0.1.17"
pyyaml = "^5.3"
pytest = "^5.3.5"
tqdm = "^4.43.0"
mlflow = "^1.7.0"
torch = "^1.4.0"
gym = "^0.17.1"
psutil = "^5.7.0"
pyprind = "^2.11.2"
scikit-learn = "^0.23.1"
pyarrow = ">=0.17.1"
gitpython = "^3.1.7"
sphinx = "^3.1.2"
sphinx-autoapi = "^1.4.0"
sphinx-rtd-theme = "^0.5.0"
sphinxcontrib-napoleon = "^0.7"
sphinxcontrib-katex = "^0.6.1"
recommonmark = "^0.6.0"
backoff = "^1.10.0"
matplotlib = "^3.3.2"
seaborn = "^0.11.0"
rlpyt = {git = "https://github.com/afansi/rlpyt.git"}
kaleido = "0.2.1"

[tool.black]
exclude = '.vscode|.eggs|.venv|.ipynb'

[tool.poetry.dev-dependencies]
isort = "^4.3"
pylama = "^7.7"
pytest-cov = "^2.8.1"
pytest-watch = "^4.2.0"
black = "^19.10b0"

[tool.semantic_release]
version_variable = "pyproject.toml:version"
upload_to_pypi = false
commit_message = "Version generated by python-semantic-release [ci skip]"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"

